refactor: extract SubsonicID3Controller — fixes #80
